+++ docs/changelog.md
@@ -5,6 +5,7 @@
 * Added `NotebookMetadataTestFile` for reading tests from Jupyter Notebook metadata
 * Added optional storage of tests in notebook metadata for Otter Assign
 * Added token argument to Otter Generate and Otter Assign
+* Removed deprecated tool Otter Service
 
 **v2.2.2:**